Only fetch mainnets on releaser page
